backend/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
Flask==3.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python @ file:///C:/Users/a878834/Downloads/mysql_connector_python-9.2.0-py2.py3-none-any.whl#sha256=d007c05a48fc076e5ac7ad3d76e332fa5b0e885853c2da80071ade2d6e01aba5
mysqlclient==2.2.7
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
fastapi
uvicorn
python-multipart
Flask
flask-cors
pyzbar
Pillow
